sm_cloud_run/requirements.txt
# Dependencias do ETL
pandas==1.4.2
numpy==1.24.4
pyjanitor==0.21.2
dbfread==2.0.7
more-itertools==10.2.0
pysus==0.6.0

pyjanitor==0.21.2
frozendict==2.1.1
uuid6==2022.6.25
roman==3.3

sqlalchemy[mypy]==1.4.27
psycopg2-binary==2.9.4
python-dotenv==0.19.2

frozenlist==1.2.0
# types-frozendict = "0.1.6"


google-cloud-storage
google-cloud-secret-manager

# Dependencias Flask
Flask==3.0.0
gunicorn==20.1.0
Werkzeug==3.0.1
waitress==3.0.0
